create study modal data schema and ui schema
